Add transpose option to expression loader
tl;dr: Small change that needs lots of restructuring.

The existing loader was spread out over multiple files making data
transforms tricky to do reliably, so the nuts and bolts of parsing has
been consolidated into TextDelimitedMatrix.java. This enabled the
addition of the option of transposing the input data. In doing so the
loading procedure needed to be generalised somewhat as the assumptions
of the input data starting on a particular row etc. were no longer
valid. Futhermore the UI for selecting the data in the input file also
needed generalisation in order to facilitate changing the row on which
the data started.