Encode the parameters in cache file name
Recently several correlation parameters have been added. This is
problematic when it comes to storing the cache files since different
parameters produce different data, thus invalidating the cache. The
change here encodes the correlation parameters in the file name so that
separate cache files are created for different sets of parameters.