Check hash of extracted libraries
It turns out that people want to be able to run several copies of
BioLayout at once. This was problematic as each copy would try and
extract its set of libraries into the same directory. Windows has a
problem with this due to file locking, so now we check if there is an
existing extracted file and whether it's correct by comparing its SHA-1
hash to the once we would potentially extract.