Handle process edge glyphs better
mEPN 2010 includes nodes of the style --<?>--> where ? is one of A, I or
C. They were used in preference to lines with exotic endings which were
not supported by yED at the time. They have since been deprecated since
yED was improved, but their use persists in some diagrams so they must
be supported.