Remove remaining JNI usage
It has been a difficult decision but the marginable speed benefit that
is attained by using JNI versions of the expression correlation and
layout algorithms is not worth the pain in terms of compilation and
maintenance. In particular running a 32bit JVM on a 64bit windows
install (the general case at Roslin), does not cooperate well with the
JNI parts of BioLayout. It's possible this could be re-examined in
future, preferably with an eye to properly leverage the advantages of
going native (intrinsics, SIMD etc.), rather than just reimplementing
the Java code in C.