Rewrite the expression file data bounds guesser
It turns out that the expression file format isn't particularly well
thought out in that the location and dimensions of the actual data is
not encoded into the format itself and must be heuristically guessed
when it's loaded. Previously assumptions were made about the numerical
content of files, and the guessing would not identify a contiguous block
of data, but rather the first numerical column and row. This is now
fixed.